requirements.txt
fastapi>=0.75.0
uvicorn>=0.11.8
python-dotenv>=0.19.2
requests>=2.22.0
requests_oauthlib>=1.3.0